Move everything into src and bring the schema down into the python package
