+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "monarch-py"
-version = "0.2.1"
+version = "0.2.2"
 description = "Monarch Initiative data access library"
 authors = [
     "Kevin Schaper <kevin@tislab.org>",
