+++ src/monarch_py/datamodels/entity.yaml
@@ -1,6 +1,6 @@
 id: https://w3id.org/monarch/monarch-py
 name: EntityIndex
-description: Data model for Monarch Solr
+description: Data datamodels for Monarch Solr
 imports:
 - linkml:types
 prefixes:
