+++ src/monarch_py/query.py
@@ -1,8 +1,6 @@
 from dataclasses import dataclass
 from urllib.parse import urlencode, urljoin
 
-from monarch_py.utilities.utils import *
-
 
 @dataclass
 class SolrQuery:
