+++ tests/integration/solr/test_association.py
@@ -1,4 +1,4 @@
-from monarch_py.implementations.solr.solr_implentation import SolrImplementation
+from src.monarch_py.implementations.solr.solr_implentation import SolrImplementation
 
 
 def test_associations():
